+++ internal/runner/testdata/pytest_pants/passing_test.py
@@ -1,2 +1,0 @@
-def test_happy():
-    assert 3 == 3